Move demo root defn into `DockerDev` config class
Since this location is fixed by the definition of the pfsc-server image, it
makes sense to define it as a constant in the `DockerDev` class, instead
of configuring it from the .env file. The latter still has a config var with
which to say whether the demo repos will actually be provided or not.