Make repo root deps override sphinx conf
Sphinx needs to know the versions at which we're taking pfsc dependencies,
so there's a var for this in your Sphinx conf.py (defined by the sphinx-proofscape extension).

However, we have a kind of nested build architecture, where a Sphinx build is being
triggered from within a surrounding PISE build. This means there's an opportunity for
double talk, which we don't want.

Since PISE is the outer build process, it should override. This means we want to read the
declared dependencies out of the repo root module, and use the `-D` switch to make these
values override any that may have been written into the Sphinx conf.py.