Make `ExampWidget.rebuildSequence()` more robust
Now, when a widget has an error, we still try to rebuild any widgets in the
sequence that are *not* descendants of this one.

To illustrate what the problem was with the existing design,
suppose B and C depend on A, but are independent of each other.
Suppose A has changed, causing an error in B, but C is still fine.
If we give up as soon as we detect the error in B, then C will still be
grayed out. And if we later update B to work with the new value of A,
then, since C is *not* a descendant of B, this update will fail to un-gray C.
C will still look disabled, even though it was always fine.

Under the new design, we note the error in B, but we continue to rebuild C.