Use resize ghost for disp widget editors
We use this instead of the `ds` scroll adjustment solution from last commit.
That was making the resizing a little bit difficult to control, since it would
suddenly accelerate when the scroll offset began to take effect.

Now we simply show a resize ghost, and do not actually adjust the size of
the editor until mouseup.