Give `DispWidget` a `preactivate()` method
This is so the `val()` method can return the newly updated code
when it is invoked by `ExampWidget.activate()`. This means now
updates show up immediately on reload of an existing notes panel
(instead of having to wait for the first rebuild via parameter choice,
as was happening up to now).