Remove examp submodule
It was considered that maybe the pfsc-examp project should become a subproject of pise;
however, it seems better that their version numbers be free to vary independently.
So, instead of making an actual examp dir, we could have used `git submodule` for this
topic branch. We now remove the directory, after adding these changes to the pfsc-examp repo
itself.