Modify active pane announcement
There is an announcement that is triggered from the `ContentManager.openContentInPane()`
method that is responsible for loading all content. We now make it so that this announcement
is not made until after the content has finished loading.

I beleive this is fine (or maybe even a slight improvement) for each of the four existing active
pane listeners we have, currently. It will be necessary for the new listener we now want to set
up, for passing doc highlights based on active panes.