Support pdf mode select via new cursors in pfsc-pdf
In pfsc-pdf v3.2.0, we will have an "Enrichments Mode" and a "Box Select Mode"
in the secondary toolbar.

The enrichments mode is where highlights from a supplier are activated.

Box select mode is as before, except now we have a per-panel control, instead
of the global one in the PDF menu we have had up to now.