Use named hl's to navigate chart --> doc
This means that instead of using ad hoc highlights, as we did before, we now are
selecting one of the "named" highlights, built into the new highlight layer.

This might mean that the `docRef` property on nodes that do make a doc ref is now
no longer needed at all. For now I leave it alone.

Note: Had to remove the `position: absolute` CSS property from the `hl-zone` class.
(Luckily, it was not actually doing anything.) This is so that pdf.js's built in 'scrollIntoView()`
functionality works, when the div we want to scroll to is one of the `hl-region` elements
making up a highlight.