Remove unused `movePaneToWindow` event
This was an old event. We no longer use it for anything.
While it seems potentially useful, I'm removing it now for
the sake of simplicity.

Theoretically, the new `ContentManager.movePaneToAnotherWindow()` method that
actually does the work here now dispatches any events that should be necessary when
a pane is moved.