Provide `NotesManager.handleDocHighlightClick()`
Now clicking highlights in a doc panel that were supplied by a notes page can navigate
the notes page.

Right now, the Alt key semantics are not quite what we wanted. We're employing the browser's
build-in focus functionality, to make widgets appear "selected". This works well, including automatically
"deselecting" when the user clicks away from them. However, it means that even when Alt is not pressed,
we still scroll the item into view, if it was currently off screen.

Might improve this at another time.