Bugfix in linking maint. for panel window move
When a panel is moved to a new window, our maintainence of linking maps was
not quite correct. A mapping `u, x |--> w`, for u the moving panel, was remaining
defined in the old window, as well as becoming defined in the new.

We make three changes to fix this:
1. Have `GlobalLinkingMap` handle 'paneMovedToAnotherWindow' *before* `ContentManager.handleClosingPaneId()` happens.
2. In `GlobalLinkingMap.noteMovedPanel()`, exclude self from the `addSet()` call.
3. When `GlobalLinkingMap` handles 'localPaneClose', do `deleteForAllX()` only for our local component, not globally.

Change (1) is necessary so that the mapping `u, x |--> w` still exists, and is available to be copied to the new window.
This then necessitates change (2), because our clever idea of "delete + add = move" no longer works, if the uuid is still going to
be present in the window from which the panel is departing.

Change (3) could simply be described as a correction of the handling of a closing panel, but the correction is only *relevant*
when a panel is being moved from one window to another, hence it's a part of this commit. When a panel closes, all windows need
to know to purge that panel *as a target*, but only the window where the panel lived needs to purge it *as a primary id*, so the latter
should be a local action. It *matters* that this action be local only when we're in the midst of moving a panel, so that new linking
records, held by the new window, are not also purged.