Revise return signature of `openContentInActiveTC`
The return value of a content pane was not being consumed anywhere.
Therefore there was no need for an auxiliary method to return just the id.
The latter was serving as an adapter, to allow this to be called over BroadcastChannel.

Now we return a promise that resolves with the pane id.
We make the method async, and it awaits opening content in a new pane (in that case)
before returning.