Bugfix in linking maint. for panel window move (again)
This time the issue was when a target (not source) panel was the one that's moving.

When the moved panel closed in its old window, we purged its uuid as a target from
all linking maps, thus unlinking it, which we did not want to do.

So now, when any panel is closing, the `GlobalLinkingMap` handler for the event checks
to see if the uuid still exists in some other window. Iff it does, then we decline to purge the
uuid as a target.