Bugfix in `PdfController.transitionToState()`
Our tests to avoid reloading a doc that had already been loaded were a little flawed.
We were missing the case where it was the same doc, but requested via a different
means of loading.

The doc could have been opened manually from the filesystem, and then a call to
`updateContent()` with a CDO naming the same doc via URL would cause us to
reload unnecessarily.

We now add a test for whether `info.docId === this.docId`. This is a better way to
detect that it's the same doc.