Make vacuum links for `L_C`
We rewrite comments in `ChartManager.makeDefaultLinks()` to describe it as applicable both
to newly opened deducs, and pre-existing ones. No change necessary to the code itself.

Then we simply call this again on a still-existing chart panel u for which some (u, x) has become
newly undefined, provided we are able to find some deduc in panel u that still references doc Id x.