Refactor subscriptions / refresh on rebuild
We've had some fixme's in the code for years, regarding refactoring
of the code in `ChartManager` and `NotesManager` that's in charge of
managing subscriptions, and the refreshing of open content when that
content has been rebuilt.

We now introduce a `SubscriptionManager` class, and each of
`ChartManager` and `NotesManager` simply instantiates and uses one
of these.