Improve clean-up code, for old highlight elements
We now make a call to unbind the supplier menu from old divs.

If we have brand new Highlights (such as when a supplier has been
rebuilt) then there's nothing to do here, but when an existing page
is re-rendered (e.g. when changing zoom level in a PDF viewer), then
we do have old divs here, from which to unbind.

I'm not sure if this is important or not, to avoid memory leak, but
it seems worth trying to clean up properly.