Redesign highlight click semantics
It seems there is no reason to click a highlight unless you want to see the
thing it points to, so click (without alt) should cause scroll.

Furthermore, we can do a nicer job with scrolling.

Therefore, new rules:

click: select and scroll. Scroll object to *middle* of panel (not top),
  following the 'distant' policy.

alt-click: select but do not scroll.