Bugfix with `.hl-multi` class
It turns out it's possible (e.g. when reloading the ISE) for the last
call to `Highlight.redoSupplierMenu()` to happen *before* the zones for this Highlight
have even been built yet. So, we're trying to set the multi class, but there are no
divs on which to set it.

As a result, clicking a shared highlight still brings up the context menu properly, but
the cursor is not the context-menu cursor we want.

Therefore we now reassess the multi class in two places, one being where we were
already doing it, in `redoSupplierMenu()`, the other being immediately after a zone
div is constructed, in `Highlight.buildZoneDiv()`.