Add automatic mapping for clone sets
The normal thing, when making a set of clones, is that you want any outgoing links from the
cloned nodes to be first resolved as usual, i.e. within their home context, but then be mapped
according to your full set of clones.

In other words, suppose nodes A, B, and C live in the original, and A has outgoing links to B
and C. If you make clones A', B' of A, B, then you want A' to link to B' and to C.

We add this "automatic mapping" behavior now.

In the future, we could support "custom mapping", by providing some syntax (maybe a "body"
section for `clone` commands) where the user can specify mappings. Internally, these could
simply be added to the new `self.relpathResolutions` dictionary of the cloned object.