Bugfix in `PfscChartWidgetBuilder`
The fix from a couple of commits ago, to try to silence warnings, was flawed.
Apparently due to some special attribute handling in the `SphinxRole` and `SphinxDirective`
base classes, we can't assign to `self.config`.

So instead we make the method static, as a way to silence warnings about
attributes `self` does not appear to have.