Remove several redundancies from the sphinx extension
There were many FIXMEs in the Sphinx extension, pointing to places
where we had copied code from pise/server, until such time as we
could properly refactor, and *reuse* that code.

In this commit we do most of that refactoring. Still remaining is
to reconcile the extension's `ChartWidget` class with the original.