Give each Sphinx page full "page data"
Every Sphinx page (whether it defines any widgets or not) now has a script tag defining
a full "page data" object of the same format as the page data for an old-style annotation page.

In this commit we also do some refactoring, including renaming the extension's chart widget
class to `SphinxChartWidget`.