Maintain `src` attr in sphinx iframes
We were not updating the iframe's `src` at all, ever, which meant that the moment
we moved the iframe (to another tab container), causing it to reload, it would become
completely blank.

We now set `this.iframe.src` *instead of* `this.cw.location` when updating the location
of a `SphinxViewer`.