Eliminate use of `...surrounding...` methods for widget clicks
NavWidgets and LinkWidgets made use of the various "get surrounding"
methods, to determine the Dijit panel in which the click happened. We now
employ other means to determine the panel, since the "surrounding" methods
won't work when the clicked element was inside an iframe (as will be the case
when the widget was in a Sphinx page).