Change widget naming rule in Sphinx pages
User-supplied names may be given at the start of the raw label,
whether in role, directive, or substitution form. They must be
a valid libseg, followed by a colon.

Widgets lacking user-supplied names are given generated names
of the form f'_w{n}'. User-supplied names are not allowed to begin
with an underscore, so there can be no collision between system-generated
and user-supplied names. This greatly simplifies the process of
generating names, in particular allowing them to be generated as
each widget is encountered, instead of having to wait until an entire
page has been processed.

Another major difference between Sphinx pages and classic annotations
is that there is no programmatic editing of the former; the generated
widget names are *not* recorded in the source file. This means that
in WIP the generated names are unstable, but of course they are stable
in a numbered release.