Repair test regarding cyclic reference
This is a nice demonstration of the fact that cyclic references between siblings defined within a module
(or within an annotation) are now possible, at least in principle.

In the present case, it's still an error because it results in an inability to compute the topological order
among a set of examp widgets. But this is a different error than we got before, when we were simply
unable to define the widgets, with cyclic references to each other.