Stop using global module cache for builds
Now the `Builder` forms its own module cache, unique to this build.
This cache is used whenever the `Builder` loads a moule, as well as
when any `PendingImport` constructed in this build process loads a
module.

We were getting bizarre, history-dependent errors when running the
whole unit test suite, which was enough additional motivation to switch
to this design, which had already been contemplated anyway.