Cascade libpaths in rst modules upon object add
In pfsc modules built by `ModuleLoader`, we call `cascadeLlibpaths()` on annos and
deducs immediately after they are constructed. We can do this because the AST traversal
here is bottom-up, so these objects already have all of their child objects at this time.

In Sphinx, the process is instead top-down. This means we cannot call `cascadeLibpaths()`
on a `SphinxPage` immediately after it is formed, because at that time it does not yet have
any of its child objects (namely, its Widget instances). So this is too early.

But we also cannot wait until the call to `SphinxPage.resolve()`, because e.g. widgets need
to already know their libpaths in order to generate their node in the doctree (their `<a>` tag),
which contains their libpath in a class name. So this is too late.

Intead, we need to pass down libpaths each time an object is added to a `SphinxPage`.