Eliminate write to `lib` dir
This brings the pfsc side into parity with the rst side, where there will be no
programmatic writes to source.

We still generate missing widget names, exactly as before. We just don't write
the updated module to the `lib` dir.

This is necessary so that file modification time can actually reflect how recently
the user changed the file, and pickling the module cache can be useful, for
speeding up rebuilds.