Revise cache policy controls
Using the `ALWAYS` policy in `PendingImport` was not correct, because for imports of
modules *outside* the repo being built, we want the `TIME` policy to apply.

However, use of `ALWAYS` was meant to guarantee use of the local modules that had
already been built during the READ phase of a build. Changing it to `TIME` broke several
unit tests, due to the `t_m + 1 >= t_r` trick in `load_module()`, confirming this motivation.

So now we try something new: `load_module()` will always use the in-mem cache, when
it's a hit; the passed `CachePolicy` arg now only controls how it uses the on-disk cache, i.e.
pickle files.