Lift Sphinx pages in build tree
To reduce clutter in the build tree, we "lift" the nodes for Sphinx pages,
meaning that they *replace* the node for the module in which they are
defined.

Other items defined in these modules are listed under the page node, as
if it were their parent.

Since you have to double-click a page node to open it, we go back to the
interaction mode where the click to expand a tree item has to be on the
"triangle" icon itself, not just anywhere in the item. This is the same as the
behavior in some other IDEs that I use, and I think it's fine.