Repair `WriteHandler` and move to `repoBuilt` event
The `WriteHandler` now reduces the given list of `buildpaths` to a set of repopaths
without repeats. Thus, it will only build any given repo once, per request.

Accordingly, it now emits a `repoBuilt` event, instead of the old `moduleBuilt` event.

We update the `SubscriptionManager` on the client side to handle the new event
instead of the old one. Now, any built product that is currently open in the ISE will
be refreshed, when the repo that defines it is rebuilt.