Correct the test in `Builder` for doing a Sphinx build
We want to do a Sphinx build if and only if there are one or more rst files
in the repo.

However, in that case we do a check to see whether the two required files
(conf.py and index.rst at the repo root level) are present, and raise an
informative exception if not.