Redesign enrichments format (again)
We no longer consolidate according to libpath.

When the same libpath is reported repeatedly by the GDB, this either
means different entities, re-using a single libpath, or that there is a
separate WIP version (now that WIP is indexed in ISOLATED mode).

In the client, the user will see the same thing as before. But actual
different entities will get their own listings, which is what we want.