Enrich widget data during module resolution
Widget data enrichment was happening during scanning; we now do it a little bit
earlier, during resolution. The reason for this is that it needs to respect the import
dependency DAG. This is because the enrichment of one widget can depend on
an imported widget having already been enriched. For example, a `PdfWidget`
can set the libpath of another in its `sel` field, meaning it wants to copy the other's
doc reference. The latter has to have been formed first, before it can be copied.