Debug autoscroll again
`BasePageViewer.goTo()` was failing to reproduce several important lines found
at the start of `PageViewer.goTo()`, where we say that, if libpath and version were not
given, then we assume that you want to stay at the same page, i.e. use the existing ones.

Along with this, we make `SphinxViewer.pageContentsUpdateStep()` more careful,
so that it loads the page only an actual page change.