Adjust widget context menus for sphinx panels
Dijit `Menu` has special code to place itself correctly inside an iframe, but you must
pass the iframe under `targetNodeIds`. You can then use `selector` to select the element(s)
on which you want to bind the menu.

Without this, the menu appears displaced from the mouse pointer by the same displacement of
the iframe's top-left corner from that of the overall window.