Repair static dir in OCA image
The built-in `css` and `img` dirs in `pise/server` were being incorrectly
put under a version dir. The URL has a version number in it, for cache busting,
but, internally, Flask routes these requests to a filesystem path that does
not include the version segment.

Due to the incorrect placement, our 404 page was failing to find the img and
css it needed.