Bugfix in mathworker and examp widgets
We were failing to handle the case in which a page with examp widgets was opened
and then closed, all before Pyodide finished loading, and the mathworker became ready.

The opening of the page puts in requests for the math worker, while its closing takes away
certain objects those requests need to work with, but does not cancel the requests themselves.

We now handle this case by carefully checking whether the objects are present, reporting
an appropriate error if not, and handling the error.