Raise meaningful excep. on PF-JSON parse err
In case an author puts malformed PF-JSON as the value of a field
in a directive-format or role-format widget in a Sphinx page, we now
raise a `PfscExcep` with a helpful error message. (We were getting an
unhelpful `LarkError`.)

This commit also resolves another, longstanding issue, by setting
  `warningiserror=True`
in our `Sphinx` builder instance. We were suffering from parse errors that were
merely printed as "warnings" that the user never saw, while working through PISE.
Now those too will be raised as `PfscExcep` and reported to the user.