Param widget: change 'default' req to 'init' opt
The 'default' field was always optional; each parameter type in pfsc-examp
is designed to choose a default value if you don't specify one.

But "default" is really not the right word. It's an *initial* value.

This also makes it easier to refer to the "default initial value" that the param
type will choose for you, if you don't specify one.