Repair `Widget.groupId`
The `groupId` property of a widget was not getting updated when a page was
rebuilt/reloaded.

This is because it was read once out of the widget info at construction time,
and never again.

We now use a method, instead of a property, for `groupId`.

Note: Tried to use a proper `get` method, but `Widget` is still a Dojo-style
class, which seems not to support this.