Make the startup sync check opt-in only
There are situations in which multiple apps are formed. In such cases,
automatically doing the sync check with cleanup actions can lead to
bad race conditions (e.g. competing with build operations).

To avoid such issues, the check + cleanup is now opt-in only, and we
currently opt in only in the app provided for unit tests in conftest.py.